# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * User Trading Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
  success?: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface OtpRequest {
  /** 10-digit mobile number */
  phone: string;
}

export interface OtpVerify {
  phone: string;
  otp: string;
}

export interface AccountInfo {
  id: number;
  name: string;
  maskedMobileNumber: string;
  isActive: boolean;
}

export interface AuthResponse {
  token: string;
  account: AccountInfo;
}

export interface Balance {
  /** @nullable */
  availableBalance: string | null;
  /** @nullable */
  totalBalance: string | null;
}

export interface Order {
  /** @nullable */
  orderId?: string | null;
  symbol: string;
  side: string;
  orderType: string;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  execQuantity?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  triggerPrice?: string | null;
  /** @nullable */
  avgExecutionPrice?: string | null;
  /** @nullable */
  executionFee?: string | null;
  /** @nullable */
  realisedPnl?: string | null;
  /** @nullable */
  reduceOnly?: boolean | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  createdAt?: string | null;
}

export interface OrderList {
  orders: Order[];
}

export interface OpenOrdersFilter {
  symbol?: string;
  limit?: number;
}

export interface ClosedOrdersFilter {
  symbol?: string;
  status?: string;
  fromTime?: number;
  toTime?: number;
  limit?: number;
}

export interface Position {
  /** @nullable */
  positionId?: string | null;
  symbol: string;
  /** @nullable */
  positionSide?: string | null;
  /** @nullable */
  leverage?: string | null;
  /** @nullable */
  positionSize?: string | null;
  /** @nullable */
  positionValue?: string | null;
  /** @nullable */
  positionMargin?: string | null;
  /** @nullable */
  maintMargin?: string | null;
  /** @nullable */
  avgEntryPrice?: string | null;
  /** @nullable */
  markPrice?: string | null;
  /** @nullable */
  lastPrice?: string | null;
  /** @nullable */
  unrealisedPnl?: string | null;
  /** @nullable */
  liquidationPrice?: string | null;
  /** @nullable */
  marginType?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface PositionList {
  positions: Position[];
}

export type OrderInputSide = typeof OrderInputSide[keyof typeof OrderInputSide];


export const OrderInputSide = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export type OrderInputOrderType = typeof OrderInputOrderType[keyof typeof OrderInputOrderType];


export const OrderInputOrderType = {
  MARKET: 'MARKET',
  LIMIT: 'LIMIT',
  STOP_MARKET: 'STOP_MARKET',
  TAKE_PROFIT_MARKET: 'TAKE_PROFIT_MARKET',
} as const;

export type OrderInputTimeInForce = typeof OrderInputTimeInForce[keyof typeof OrderInputTimeInForce];


export const OrderInputTimeInForce = {
  GTC: 'GTC',
  IOC: 'IOC',
  FOK: 'FOK',
} as const;

export interface OrderInput {
  symbol: string;
  side: OrderInputSide;
  order_type: OrderInputOrderType;
  quantity: number;
  price?: number;
  triggerPrice?: number;
  reduceOnly?: boolean;
  timeInForce?: OrderInputTimeInForce;
  clientOrderId?: string;
}

export interface OrderPlaced {
  orderId: string;
  status: string;
}

export interface PnlSummary {
  todayPnl: number;
  weekPnl: number;
  monthPnl: number;
  totalPnl: number;
  tradeCount: number;
}