# lib/api-zod/src/generated/types/accountInfo.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * User Trading Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface AccountInfo {
  id: number;
  name: string;
  maskedMobileNumber: string;
  isActive: boolean;
}