# lib/api-zod/src/generated/types/balance.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * User Trading Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Balance {
  /** @nullable */
  availableBalance: string | null;
  /** @nullable */
  totalBalance: string | null;
}