# lib/api-zod/src/generated/types/openOrdersFilter.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * User Trading Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenOrdersFilter {
  symbol?: string;
  limit?: number;
}