# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * User Trading Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  /** @nullable */
  orderId?: string | null;
  symbol: string;
  side: string;
  orderType: string;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  execQuantity?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  triggerPrice?: string | null;
  /** @nullable */
  avgExecutionPrice?: string | null;
  /** @nullable */
  executionFee?: string | null;
  /** @nullable */
  realisedPnl?: string | null;
  /** @nullable */
  reduceOnly?: boolean | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  createdAt?: string | null;
}