# lib/api-zod/src/generated/types/position.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * User Trading Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Position {
  /** @nullable */
  positionId?: string | null;
  symbol: string;
  /** @nullable */
  positionSide?: string | null;
  /** @nullable */
  leverage?: string | null;
  /** @nullable */
  positionSize?: string | null;
  /** @nullable */
  positionValue?: string | null;
  /** @nullable */
  positionMargin?: string | null;
  /** @nullable */
  maintMargin?: string | null;
  /** @nullable */
  avgEntryPrice?: string | null;
  /** @nullable */
  markPrice?: string | null;
  /** @nullable */
  lastPrice?: string | null;
  /** @nullable */
  unrealisedPnl?: string | null;
  /** @nullable */
  liquidationPrice?: string | null;
  /** @nullable */
  marginType?: string | null;
  /** @nullable */
  status?: string | null;
}